create Mojito.Pool behaviour and Mojito.Pool.Poolboy implementation
