+++ test/mojito_test.exs
@@ -171,7 +171,7 @@ defmodule MojitoTest do
     end
 
     it "accepts pool: pid" do
-      child_spec = Mojito.Pool.Single.child_spec()
+      child_spec = Mojito.Pool.Poolboy.Single.child_spec()
       {:ok, pool_pid} = Supervisor.start_child(Mojito.Supervisor, child_spec)
 
       assert(
