+++ test/pool/poolboy/manager_test.exs
@@ -1,11 +1,11 @@
-defmodule Mojito.Pool.ManagerTest do
+defmodule Mojito.Pool.Poolboy.ManagerTest do
   use ExSpec, async: true
 
   context "calls" do
     it "implements get_pools" do
       assert(
         [] =
-          GenServer.call(Mojito.Pool.Manager, {:get_pools, {"example.com", 80}})
+          GenServer.call(Mojito.Pool.Poolboy.Manager, {:get_pools, {"example.com", 80}})
       )
     end
 
@@ -13,7 +13,7 @@ defmodule Mojito.Pool.ManagerTest do
       assert(
         [] =
           GenServer.call(
-            Mojito.Pool.Manager,
+            Mojito.Pool.Poolboy.Manager,
             {:get_pool_states, {"example.com", 80}}
           )
       )
